+++ packages/alfa-aria/src/name/index.ts
@@ -1,2 +1,4 @@
 export * from "./name";
 export * from "./predicate";
+export * from "./source";
+export * from "./state";
