+++ packages/alfa-aria/src/name/name.ts
@@ -18,6 +18,8 @@ import { Feature } from "../feature";
 import { Role } from "../role";
 
 import * as predicate from "./predicate";
+import { Source } from "./source";
+import { State } from "./state";
 
 import { isProgrammaticallyHidden } from "../dom/predicate/is-programmatically-hidden";
 
@@ -32,14 +34,14 @@ const { getElementIdMap } = Query;
  * @public
  */
 export class Name implements Equatable, Serializable<Name.JSON> {
-  public static of(value: string, sources: Iterable<Name.Source> = []): Name {
+  public static of(value: string, sources: Iterable<Source> = []): Name {
     return new Name(value, Array.from(sources));
   }
 
   private readonly _value: string;
-  private readonly _sources: Array<Name.Source>;
+  private readonly _sources: Array<Source>;
 
-  private constructor(value: string, sources: Array<Name.Source>) {
+  private constructor(value: string, sources: Array<Source>) {
     this._value = value;
     this._sources = sources;
   }
@@ -48,7 +50,7 @@ export class Name implements Equatable, Serializable<Name.JSON> {
     return this._value;
   }
 
-  public get source(): ReadonlyArray<Name.Source> {
+  public get source(): ReadonlyArray<Source> {
     return this._sources;
   }
 
@@ -91,486 +93,6 @@ export namespace Name {
     [key: string]: json.JSON;
     value: string;
     sources: Array<Source.JSON>;
-  }
-
-  export type Source =
-    | Source.Data
-    | Source.Descendant
-    | Source.Ancestor
-    | Source.Label
-    | Source.Reference;
-
-  export namespace Source {
-    export type JSON =
-      | Data.JSON
-      | Descendant.JSON
-      | Ancestor.JSON
-      | Label.JSON
-      | Reference.JSON;
-
-    export class Data implements Equatable, Serializable<Data.JSON> {
-      public static of(text: Text): Data {
-        return new Data(text);
-      }
-
-      private readonly _text: Text;
-
-      private constructor(text: Text) {
-        this._text = text;
-      }
-
-      public get type(): "data" {
-        return "data";
-      }
-
-      public get text(): Text {
-        return this._text;
-      }
-
-      public equals(value: unknown): value is this {
-        return value instanceof Data && value._text.equals(this._text);
-      }
-
-      public *[Symbol.iterator](): Iterator<Node> {
-        yield this._text;
-      }
-
-      public toJSON(): Data.JSON {
-        return {
-          type: "data",
-          text: this._text.path(),
-        };
-      }
-    }
-
-    export namespace Data {
-      export interface JSON {
-        [key: string]: json.JSON;
-        type: "data";
-        text: string;
-      }
-    }
-
-    export function data(text: Text): Data {
-      return Data.of(text);
-    }
-
-    export class Descendant
-      implements Equatable, Serializable<Descendant.JSON>
-    {
-      public static of(element: Element, name: Name): Descendant {
-        return new Descendant(element, name);
-      }
-
-      private readonly _element: Element;
-      private readonly _name: Name;
-
-      private constructor(element: Element, name: Name) {
-        this._element = element;
-        this._name = name;
-      }
-
-      public get type(): "descendants" {
-        return "descendants";
-      }
-
-      public get element(): Element {
-        return this._element;
-      }
-
-      public get name(): Name {
-        return this._name;
-      }
-
-      public equals(value: unknown): value is this {
-        return (
-          value instanceof Descendant &&
-          value._element.equals(this._element) &&
-          value._name.equals(this._name)
-        );
-      }
-
-      public *[Symbol.iterator](): Iterator<Node> {
-        yield this._element;
-        yield* this._name.sourceNodes();
-      }
-
-      public toJSON(): Descendant.JSON {
-        return {
-          type: "descendant",
-          element: this._element.path(Node.flatTree),
-          name: this._name.toJSON(),
-        };
-      }
-    }
-
-    export namespace Descendant {
-      export interface JSON {
-        [key: string]: json.JSON;
-        type: "descendant";
-        element: string;
-        name: Name.JSON;
-      }
-    }
-
-    export function descendant(element: Element, name: Name): Descendant {
-      return Descendant.of(element, name);
-    }
-
-    export class Ancestor implements Equatable, Serializable<Ancestor.JSON> {
-      public static of(element: Element, name: Name): Ancestor {
-        return new Ancestor(element, name);
-      }
-
-      private readonly _element: Element;
-      private readonly _name: Name;
-
-      private constructor(element: Element, name: Name) {
-        this._element = element;
-        this._name = name;
-      }
-
-      public get type(): "ancestor" {
-        return "ancestor";
-      }
-
-      public get element(): Element {
-        return this._element;
-      }
-
-      public get name(): Name {
-        return this._name;
-      }
-
-      public equals(value: unknown): value is this {
-        return (
-          value instanceof Ancestor &&
-          value._element.equals(this._element) &&
-          value._name.equals(this._name)
-        );
-      }
-
-      public *[Symbol.iterator](): Iterator<Node> {
-        yield this._element;
-        yield* this._name.sourceNodes();
-      }
-
-      public toJSON(): Ancestor.JSON {
-        return {
-          type: "ancestor",
-          element: this._element.path(),
-          name: this._name.toJSON(),
-        };
-      }
-    }
-
-    export namespace Ancestor {
-      export interface JSON {
-        [key: string]: json.JSON;
-        type: "ancestor";
-        element: string;
-        name: Name.JSON;
-      }
-    }
-
-    export function ancestor(element: Element, name: Name): Ancestor {
-      return Ancestor.of(element, name);
-    }
-
-    export class Label implements Equatable, Serializable<Label.JSON> {
-      public static of(attribute: Attribute): Label {
-        return new Label(attribute);
-      }
-
-      private readonly _attribute: Attribute;
-
-      private constructor(attribute: Attribute) {
-        this._attribute = attribute;
-      }
-
-      public get type(): "label" {
-        return "label";
-      }
-
-      public get attribute(): Attribute {
-        return this._attribute;
-      }
-
-      public equals(value: unknown): value is this {
-        return (
-          value instanceof Label && value._attribute.equals(this._attribute)
-        );
-      }
-
-      public *[Symbol.iterator](): Iterator<Node> {
-        yield this._attribute;
-      }
-
-      public toJSON(): Label.JSON {
-        return {
-          type: "label",
-          attribute: this._attribute.path(),
-        };
-      }
-    }
-
-    export namespace Label {
-      export interface JSON {
-        [key: string]: json.JSON;
-        type: "label";
-        attribute: string;
-      }
-    }
-
-    export function label(attribute: Attribute): Label {
-      return Label.of(attribute);
-    }
-
-    export class Reference implements Equatable, Serializable<Reference.JSON> {
-      public static of(attribute: Attribute, name: Name): Reference {
-        return new Reference(attribute, name);
-      }
-
-      private readonly _attribute: Attribute;
-      private readonly _name: Name;
-
-      private constructor(attribute: Attribute, name: Name) {
-        this._attribute = attribute;
-        this._name = name;
-      }
-
-      public get type(): "reference" {
-        return "reference";
-      }
-
-      public get attribute(): Attribute {
-        return this._attribute;
-      }
-
-      public get name(): Name {
-        return this._name;
-      }
-
-      public equals(value: unknown): value is this {
-        return (
-          value instanceof Reference && value._attribute.equals(this._attribute)
-        );
-      }
-
-      public *[Symbol.iterator](): Iterator<Node> {
-        yield this._attribute;
-        yield* this._name.sourceNodes();
-      }
-
-      public toJSON(): Reference.JSON {
-        return {
-          type: "reference",
-          attribute: this._attribute.path(),
-          name: this._name.toJSON(),
-        };
-      }
-    }
-
-    export namespace Reference {
-      export interface JSON {
-        [key: string]: json.JSON;
-        type: "reference";
-        attribute: string;
-        name: Name.JSON;
-      }
-    }
-
-    export function reference(attribute: Attribute, name: Name): Reference {
-      return Reference.of(attribute, name);
-    }
-  }
-
-  /**
-   * @internal
-   */
-  export class State implements Equatable, Serializable<State.JSON> {
-    private static _empty = new State([], None, None, false, false);
-
-    public static empty(): State {
-      return this._empty;
-    }
-
-    private readonly _visited: Array<Element>;
-    // which element has an aria-labelledby causing the current traversal?
-    private readonly _referrer: Option<Element>;
-    // which element was the target of aria-labelledby?
-    private readonly _referred: Option<Element>;
-    private readonly _isRecursing: boolean;
-    private readonly _isDescending: boolean;
-
-    private constructor(
-      visited: Array<Element>,
-      referrer: Option<Element>,
-      referred: Option<Element>,
-      isRecursing: boolean,
-      isDescending: boolean,
-    ) {
-      this._visited = visited;
-      this._referrer = referrer;
-      this._referred = referred;
-      this._isRecursing = isRecursing;
-      this._isDescending = isDescending;
-    }
-
-    /**
-     * The elements that have been seen by the name computation so far. This is
-     * used for detecting circular references resulting from things such as the
-     * `aria-labelledby` attribute and form controls that get their name from
-     * a containing `<label>` element.
-     */
-    public get visited(): Iterable<Element> {
-      return this._visited;
-    }
-
-    /**
-     * The element that referenced the name computation.
-     * (this is the element on which aria-labelledby is set)
-     */
-    public get referrer(): Option<Element> {
-      return this._referrer;
-    }
-
-    /**
-     * The element that is referenced during the name computation.
-     * (this is the target of the aria-labelledby attribute)
-     */
-    public get referred(): Option<Element> {
-      return this._referred;
-    }
-
-    /**
-     * Whether or not the name computation is the result of recursion.
-     */
-    public get isRecursing(): boolean {
-      return this._isRecursing;
-    }
-
-    /**
-     * Whether or not the name computation is the result of a reference.
-     */
-    public get isReferencing(): boolean {
-      return this._referrer.isSome();
-    }
-
-    /**
-     * Whether or not the name computation is descending into a subtree.
-     */
-    public get isDescending(): boolean {
-      return this._isDescending;
-    }
-
-    public hasVisited(element: Element): boolean {
-      return this._visited.includes(element);
-    }
-
-    public visit(element: Element): State {
-      if (this._visited.includes(element)) {
-        return this;
-      }
-
-      return new State(
-        [...this._visited, element],
-        this._referrer,
-        this._referred,
-        this._isRecursing,
-        this._isDescending,
-      );
-    }
-
-    public recurse(isRecursing: boolean): State {
-      if (this._isRecursing === isRecursing) {
-        return this;
-      }
-
-      return new State(
-        this._visited,
-        this._referrer,
-        this._referred,
-        isRecursing,
-        this._isDescending,
-      );
-    }
-
-    /**
-     * @remarks
-     * This set both _referrer and _referred, so that they will always be
-     * either both Some or both None.
-     *
-     * @remarks
-     * We currently have no way to clear references since we currently have no
-     * use for it.
-     */
-    public reference(referrer: Element, referred: Element): State {
-      if (
-        this._referrer.includes(referrer) &&
-        this._referred.includes(referred)
-      ) {
-        return this;
-      }
-
-      return new State(
-        this._visited,
-        Option.of(referrer),
-        Option.of(referred),
-        this._isRecursing,
-        this._isDescending,
-      );
-    }
-
-    public descend(isDescending: boolean): State {
-      if (this._isDescending === isDescending) {
-        return this;
-      }
-
-      return new State(
-        this._visited,
-        this._referrer,
-        this._referred,
-        this._isRecursing,
-        isDescending,
-      );
-    }
-
-    public equals(state: State): boolean;
-
-    public equals(value: unknown): value is this;
-
-    public equals(value: unknown): boolean {
-      return (
-        value instanceof State &&
-        Array.equals(value._visited, this._visited) &&
-        value._referrer.equals(this._referrer) &&
-        value._referred.equals(this._referred) &&
-        value._isRecursing === this._isRecursing &&
-        value._isDescending === this._isDescending
-      );
-    }
-
-    public toJSON(): State.JSON {
-      return {
-        visited: this._visited.map((element) => element.path()),
-        referrer: this._referrer.map((element) => element.path()).getOr(null),
-        referred: this._referred.map((element) => element.path()).getOr(null),
-        isRecursing: this._isRecursing,
-        isDescending: this._isDescending,
-      };
-    }
-  }
-
-  export namespace State {
-    export interface JSON {
-      [key: string]: json.JSON;
-      visited: Array<string>;
-      referrer: string | null;
-      referred: string | null;
-      isRecursing: boolean;
-      isDescending: boolean;
-    }
   }
 
   export function from(node: Element | Text, device: Device): Option<Name> {
@@ -948,11 +470,4 @@ export namespace Name {
 
 function flatten(string: string): string {
   return string.replace(/\s+/g, " ");
-}
-
-/**
- *
- */
-function normalize(value: string): string {
-  return value.trim().replace(/\s+/g, " ");
 }
